requirements.txt
certifi==2025.8.3
charset-normalizer==3.4.3
contourpy==1.3.3
cycler==0.12.1
filelock==3.13.1
fonttools==4.60.0
fsspec==2024.6.1
idna==3.10
ImageHash==4.3.2
Jinja2==3.1.4
kiwisolver==1.4.9
lap==0.5.12
MarkupSafe==2.1.5
matplotlib==3.10.6
mpmath==1.3.0
networkx==3.3
numpy==2.1.2
opencv-python==4.12.0.88
packaging==25.0
pillow==11.0.0
polars==1.33.1
psutil==7.1.0
pyparsing==3.2.5
python-dateutil==2.9.0.post0
PyWavelets==1.9.0
PyYAML==6.0.2
requests==2.32.5
scipy==1.16.2
setuptools==70.2.0
six==1.17.0
sympy==1.13.1
torch==2.5.1+cu121
torchaudio==2.5.1+cu121
torchvision==0.20.1+cu121
typing_extensions==4.12.2
ultralytics==8.3.203
ultralytics-thop==2.0.17
urllib3==2.5.0
